fix: [WLEO-725] Add dPoP support to FetchTokenResponseOptions (#49)
#### List of Changes
Added support for DPoP (Demonstration of Proof-of-Possession) by
introducing a new property in `FetchTokenResponseOptions` and updating
the request headers accordingly.

#### Motivation and Context
This change is required to enhance security in token requests by
allowing the inclusion of a DPoP proof, which helps verify the
legitimacy of the token request.

#### How Has This Been Tested?
Changes were tested by updating existing unit tests to include scenarios
for DPoP proof handling and verifying that the correct headers are sent
in the token request.